Added basic AI, basement and overworld
Ai has a state machine for patrolling, chasing the player, idling near the player and running back when the player is too close. They fire when they can see and are chasing the player.
Dungeon/basement has been started with a few rooms, room changes and basic load/unload of assets (enemies and projectiles so far). 
Overworld has three buildings and a moving NPC (as long as I didn't break that during the day)

Assets are also non-blurry anymore.